cryptomarket/requirements.txt
aniso8601==9.0.1
backports.zoneinfo==0.2.1
click==8.1.2
coinapi.rest.v1==1.3
Flask==2.1.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
marshmallow==3.13.0
peewee==3.14.10
peewee-migrate==1.4.7
psycopg2-binary==2.9.1
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2022.1
six==1.16.0
waitress==2.0.0
Werkzeug==2.1.1
zipp==3.8.0
